moves creation of default attach rules from config to event handlers
adjusts config tests accordingly
