+++ internal/action/common.go
@@ -3,10 +3,13 @@ package action
 import (
 	"github.com/keptn-contrib/dynatrace-service/internal/adapter"
 	"github.com/keptn-contrib/dynatrace-service/internal/common"
+	"github.com/keptn-contrib/dynatrace-service/internal/dynatrace"
 )
 
 const eventSource = "Keptn dynatrace-service"
 const bridgeURLKey = "Keptns Bridge"
+
+const contextless = "CONTEXTLESS"
 
 func createCustomProperties(a adapter.EventContentAdapter, imageAndTag common.ImageAndTag, bridgeURL string) map[string]string {
 	customProperties := map[string]string{
@@ -38,4 +41,31 @@ func getValueFromLabels(a adapter.EventContentAdapter, key string, defaultValue 
 		return v
 	}
 	return defaultValue
+}
+
+func createDefaultAttachRules(a adapter.EventContentAdapter) *dynatrace.AttachRules {
+	return &dynatrace.AttachRules{
+		TagRule: []dynatrace.TagRule{
+			{
+				MeTypes: []string{"SERVICE"},
+				Tags: []dynatrace.TagEntry{
+					{
+						Context: contextless,
+						Key:     "keptn_project",
+						Value:   a.GetProject(),
+					},
+					{
+						Context: contextless,
+						Key:     "keptn_stage",
+						Value:   a.GetStage(),
+					},
+					{
+						Context: contextless,
+						Key:     "keptn_service",
+						Value:   a.GetService(),
+					},
+				},
+			},
+		},
+	}
 }
