+++ internal/action/deployment_finished_event_handler.go
@@ -26,8 +26,12 @@ func NewDeploymentFinishedEventHandler(event DeploymentFinishedAdapterInterface,
 }
 
 // HandleEvent handles a deployment finished event.
-func (eh *DeploymentFinishedEventHandler) HandleEvent(workCtx context.Context, replyCtx context.Context) error {
+func (eh *DeploymentFinishedEventHandler) HandleEvent(workCtx context.Context, _ context.Context) error {
 	imageAndTag := eh.eClient.GetImageAndTag(eh.event)
+
+	if eh.attachRules == nil {
+		eh.attachRules = createDefaultAttachRules(eh.event)
+	}
 
 	deploymentEvent := dynatrace.DeploymentEvent{
 		EventType:         dynatrace.DeploymentEventType,
